+++ TOneV2/Code/TOneV2/TOne.WhS.RouteSync.Entities/RouteEntities/RouteOption.cs
@@ -6,7 +6,7 @@ namespace TOne.WhS.RouteSync.Entities
     {
         public string SupplierId { get; set; }
         public decimal? SupplierRate { get; set; }
-        public Decimal? Percentage { get; set; }
+        public int? Percentage { get; set; }
         public bool IsBlocked { get; set; }
         public int NumberOfTries { get; set; }
         public bool IsValid { get { return NumberOfTries > 0; } }
