+++ TOneV2/Code/TOneV2/TOne.WhS.Routing.Entities/RouteRuleSettings/IRouteOptionPercentageExecutionContext.cs
@@ -8,7 +8,7 @@ namespace TOne.WhS.Routing.Entities
 {
     public interface IRouteOptionPercentageTarget
     {
-        Decimal? Percentage { get; set; }
+        int? Percentage { get; set; }
     }
 
     public interface IRouteOptionPercentageExecutionContext
