+++ TOneV2/Code/TOneV2/TOne.WhS.Routing.Entities/Routes/RPRouteOptionSupplier.cs
@@ -35,7 +35,7 @@ namespace TOne.WhS.Routing.Entities
         public int NumberOfBlockedZones { get; set; }
         public int NumberOfUnblockedZones { get; set; }
 
-        public decimal? Percentage { get; set; }
+        public int? Percentage { get; set; }
 
         public int SupplierServiceWeight { get; set; }
     }
